Добавил автоформатирование кода .clang-format, clangd

+++ HW5/src/B1.c
@@ -1,6 +1,6 @@
 /**
  * ДЗ-5. Си базовый уровень. гр.Д01-134 Попов. В.Г
- * B1: Ввести натуральное число вывести квадраты и кубы 
+ * B1: Ввести натуральное число вывести квадраты и кубы
  *     всех чисел от 1 до этого числа. Число не превосходит 100.
  */
 #include <stdio.h>
@@ -11,11 +11,14 @@
 int main (void)
 {
     uint16_t enter_number = 0;
-    scanf("%"SCNu16, &enter_number);
+    scanf("%" SCNu16, &enter_number);
 
-    for (uint16_t i = 1; i <= enter_number; ++i)
+    for (uint32_t i = 1; i <= enter_number; ++i)
     {
-        printf("%" PRIu32 " " "%" PRIu32 " " "%" PRIu32 "\n", i, (i * i), (i * i * i));
+        printf("%" PRIu32 " "
+               "%" PRIu32 " "
+               "%" PRIu32 "\n",
+               i, (i * i), (i * i * i));
     }
 
     return EXIT_SUCCESS;
